Moving types around and fixing styles
