+++ frontend/src/components/Version.tsx
@@ -1,11 +1,11 @@
-import "../styles/Version.css"
+import "../styles/Version.css";
 
 const Version = () => {
-  return (
-    <div className="version">
-    <h3>Version 1.0.0</h3>
-  </div>
-  )
-}
+    return (
+        <div className="version">
+            <h3>Version 1.0.0</h3>
+        </div>
+    );
+};
 
-export default Version
+export default Version;
