+++ frontend/src/styles/Account.css
@@ -1,0 +1,6 @@
+.account {
+    min-height: 100%;
+    display: flex;
+    flex-direction: column;
+    overflow-x: hidden;
+}
