+++ frontend/src/styles/AccountDefault.css
@@ -1,12 +1,6 @@
-body * {
-    overflow: hidden;
-}
-
 .account-default {
     display: flex;
     width: 100%;
-    min-height: 400px;
-    height: 100%;
     flex-direction: row;
     padding: 28px 0;
     gap: 28px;
@@ -14,4 +8,5 @@ body * {
     align-items: flex-start;
     background-color: var(--rsblack);
     color: var(--rswhite);
+    flex-grow: 1;
 }
