+++ frontend/src/styles/Version.css
@@ -1,6 +1,7 @@
 .version {
     position: absolute;
     top: 10px;
-    right: 10px;
+    right: 30px;
     color: var(--rswhite);
+    font-size: 12px;
   }
