+++ frontend/src/types/Skills.tsx
@@ -1,0 +1,31 @@
+export enum Skills {
+    Crafting = "Crafting",
+    Summoning = "Summoning",
+    Cooking = "Cooking",
+    Magic = "Magic",
+    Prayer = "Prayer",
+    Mining = "Mining",
+    Firemaking = "Firemaking",
+    Ranged = "Ranged",
+    Smithing = "Smithing",
+    Constitution = "Constitution",
+    Woodcutting = "Woodcutting",
+    Hunter = "Hunter",
+    Fishing = "Fishing",
+    Slayer = "Slayer",
+    Defence = "Defence",
+    Agility = "Agility",
+    Strength = "Strength",
+    Attack = "Attack",
+    Fletching = "Fletching",
+    Divination = "Divination",
+    Runecrafting = "Runecrafting",
+    Herblore = "Herblore",
+    Dungeoneering = "Dungeoneering",
+    Construction = "Construction",
+    Thieving = "Thieving",
+    Farming = "Farming",
+    Archaeology = "Archaeology",
+    Necromancy = "Necromancy",
+    Invention = "Invention",
+}
